+++ saleor/graphql/attribute/schema.py
@@ -7,7 +7,12 @@ from ..core.descriptions import ADDED_IN_310, ADDED_IN_311, PREVIEW_FEATURE
 from ..core.doc_category import DOC_CATEGORY_ATTRIBUTES
 from ..core.fields import BaseField, FilterConnectionField
 from ..core.utils.resolvers import resolve_by_global_id_slug_or_ext_ref
-from ..translations.mutations import AttributeTranslate, AttributeValueTranslate
+from ..translations.mutations import (
+    AttributeBulkTranslate,
+    AttributeTranslate,
+    AttributeValueBulkTranslate,
+    AttributeValueTranslate,
+)
 from .bulk_mutations import AttributeBulkDelete, AttributeValueBulkDelete
 from .filters import AttributeFilterInput, AttributeWhereInput, filter_attribute_search
 from .mutations import (
@@ -73,6 +78,7 @@ class AttributeMutations(graphene.ObjectType):
     attribute_delete = AttributeDelete.Field()
     attribute_update = AttributeUpdate.Field()
     attribute_translate = AttributeTranslate.Field()
+    attribute_bulk_translate = AttributeBulkTranslate.Field()
     attribute_bulk_delete = AttributeBulkDelete.Field()
     attribute_value_bulk_delete = AttributeValueBulkDelete.Field()
 
@@ -80,5 +86,6 @@ class AttributeMutations(graphene.ObjectType):
     attribute_value_create = AttributeValueCreate.Field()
     attribute_value_delete = AttributeValueDelete.Field()
     attribute_value_update = AttributeValueUpdate.Field()
+    attribute_value_bulk_translate = AttributeValueBulkTranslate.Field()
     attribute_value_translate = AttributeValueTranslate.Field()
     attribute_reorder_values = AttributeReorderValues.Field()