+++ saleor/graphql/core/types/common.py
@@ -36,6 +36,8 @@ from ..enums import (
     AccountErrorCode,
     AppErrorCode,
     AttributeErrorCode,
+    AttributeTranslateErrorCode,
+    AttributeValueTranslateErrorCode,
     ChannelErrorCode,
     CheckoutErrorCode,
     CollectionErrorCode,
@@ -711,9 +713,23 @@ class TranslationError(Error):
     code = TranslationErrorCode(description="The error code.", required=True)
 
 
+class TranslationBulkError(BulkError):
+    code = TranslationErrorCode(description="The error code.", required=True)
+
+
 class SeoInput(graphene.InputObjectType):
     title = graphene.String(description="SEO title.")
     description = graphene.String(description="SEO description.")
+
+
+class AttributeBulkTranslateError(BulkError):
+    code = AttributeTranslateErrorCode(description="The error code.", required=True)
+
+
+class AttributeValueBulkTranslateError(BulkError):
+    code = AttributeValueTranslateErrorCode(
+        description="The error code.", required=True
+    )
 
 
 class Weight(graphene.ObjectType):